Create Simulator parent for all simulators
