+++ src/edu/wesimulated/firstapp/simulation/PersonRolSimulator.java
@@ -9,9 +9,8 @@ import com.wesimulated.simulationmotor.des.OperationBasedExecutor;
 import edu.wesimulated.firstapp.simulation.domain.Person;
 import edu.wesimulated.firstapp.simulation.domain.Task;
 
-public class PersonRolSimulator {
+public class PersonRolSimulator extends Simulator {
 
-	private OperationBasedExecutor executor;
 	private Collection<Task> tasks;
 	private Person person;
 
@@ -20,17 +19,6 @@ public class PersonRolSimulator {
 		this.person = person;
 	}
 
-	public void startExecution() {
-		this.executor.run();
-	}
-
-	public OperationBasedExecutor getExecutor() {
-		return executor;
-	}
-
-	public void setExecutor(OperationBasedExecutor executor) {
-		this.executor = executor;
-	}
 	public void assignTask(Task task) {
 		if (this.tasks == null) {
 			this.tasks = new LinkedList<>();
@@ -40,9 +28,5 @@ public class PersonRolSimulator {
 
 	public void addBOperation(BOperation operation) {
 		this.getExecutor().addBOperation(operation);
-	}
-
-	public boolean isRunning() {
-		return !this.getExecutor().isSimulationEnd();
 	}
 }
