+++ src/edu/wesimulated/firstapp/simulation/StartEvent.java
@@ -4,7 +4,7 @@ package edu.wesimulated.firstapp.simulation;
 public class StartEvent extends SimulationEvent {
 
 	@Override
-	public void updateSimulation(PersonRolSimulator personRolSimulator, AbstractFederate personFederate) {
+	public void updateSimulation(Simulator personRolSimulator, AbstractFederate personFederate) {
 		if (personRolSimulator != null) {
 			personRolSimulator.startExecution();
 		}
